Created vecmem::details::deleter.
It's a copy-paste of vecmem::array<...>::deleter, which is now
replaced by this new type.

This became necessary as I would now also like to use the same
functionality outside of vecmem::array as well.
